Started News Page
Started news page but the backend needs to be made in firebase. For now I just used the same ones that events uses but I think it needs to be reconfigured a little bit to work for the news page. There's some more stuff to add but hopefully this is a good start.